Analysis of SportVU2.0 tracking quality
In this repository you will find the entirety of the content of my data science internship.

I was provided several csv files describing the tracking quality of the Sportvu 2.0 system. From which I had to create a predictive model to try to know it's tracking quality according to different factors.

I started by cleaning the provided data. Doing an exploratory data Analysis using pandas and numpy, from which I created different visuals with Scikit learn which I presented to my manager. Finally I created and trained a neural Network with Keras and Tensorflow to create a predictive model for the SportVU performance.